made the website locally

+++ routes/campgrounds.js
@@ -3,7 +3,7 @@ const router = express.Router();
 const Expresserror = require('../utils/ExpressError');
 const catchAsync = require('../utils/catchAsync');
 const Campground = require("../models/campground");
-const {isLoggedIn} = require('../middleware');
+const { isLoggedIn } = require('../middleware');
 
 //showig all the cities
 router.get("/", async (req, res) => {
@@ -20,6 +20,7 @@ router.get("/", async (req, res) => {
     // res.send(req.body) 
       
       const campground = new Campground(req.body.campground);
+      campground.owner = req.user._id;
       await campground.save();
       // req.flash('success','Added a new campground');
       res.redirect(`/campgrounds/${campground._id}`);
@@ -28,23 +29,45 @@ router.get("/", async (req, res) => {
   
   //detailed view of city
   router.get("/:id",async (req, res) => {
-    const campground = await Campground.findById(req.params.id).populate('reviews');
+    const campground = await Campground.findById(req.params.id).populate('reviews').populate('owner');
+    
+    console.log(campground);
     res.render("campgrounds/show", { campground });
   });
   
   router.get("/:id/edit", async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
+    
+    if(!campground.owner.equals(req.user._id)){
+      console.log("No permsission");
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    if(!campground){
+      return res.redirect('/campgrounds');
+    }
     res.render("campgrounds/edit", { campground });
   });
   router.put("/:id", async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, {
+    const campground = await Campground.findById(id);
+    if(!campground.owner.equals(req.user._id)){
+      console.log("No permsission");
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    const camp = await Campground.findByIdAndUpdate(id, {
       ...req.body.campground,
     });
     res.redirect(`/campgrounds/${campground._id}`);
   });
   router.delete("/:id", async (req, res) => {
+    
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if(!campground.owner.equals(req.user._id)){
+      console.log("No permsission");
+      return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     res.redirect("/campgrounds");
   });
